move checks to modules
